+++ .github/workflows/release.yml
@@ -4,7 +4,7 @@
 name: Release
 
 on:
-  push:
+  pull:
     branches:
       - main
