+++ terraform/state.tf
@@ -1,9 +1,8 @@
 terraform {
   backend "s3" {
-    bucket  = "radius-tracker-state"
-    key     = "tracker-state"
-    profile = "radius-tracker"
-    region  = "us-east-2"
+    bucket = "radius-tracker-state"
+    key    = "tracker-state"
+    region = "us-east-2"
   }
 }
